+++ src/imio/smartweb/core/contents/sections/external_content/views.py
@@ -3,30 +3,62 @@
 from embeddify import Embedder
 from embeddify import Plugin
 from imio.smartweb.core.contents.sections.views import SectionView
+from imio.smartweb.core.interfaces import IArcgisViewUtils
 from imio.smartweb.locales import SmartwebMessageFactory as _
 from json.decoder import JSONDecodeError
+from plone import api
+from plone.app.contenttypes.behaviors.leadimage import ILeadImage
+from Products.Five.browser import BrowserView
+from zope.i18n import translate
+from zope.interface import implementer
 
 import json
 
 
 class ExternalContentView(SectionView):
     def get_embed_external_content(self, width="100%", height=600):
+        current_lang = api.portal.get_current_language()[:2]
         plugins = [
             EaglebePlugin(),
             EllohaPlugin(),
             CognitoformPlugin(),
+            ArcgisPlugin(),
             UnknowServicePlugin(),
         ]
         extra_params = self.context.external_content_params
         plugin_config = {
             "eaglebeplugin": {"width": width},
-            "ellohaplugin": {"width": width, "extra_params": extra_params},
-            "cognitoformplugin": {"width": width, "extra_params": extra_params},
-            "unknowserviceplugin": {"width": width},
+            "ellohaplugin": {
+                "width": width,
+                "current_lang": current_lang,
+                "extra_params": extra_params,
+            },
+            "cognitoformplugin": {
+                "width": width,
+                "current_lang": current_lang,
+                "extra_params": extra_params,
+            },
+            "arcgisplugin": {
+                "width": width,
+                "url": self.context.absolute_url(),
+                "current_lang": current_lang,
+                "extra_params": extra_params,
+            },
+            "unknowserviceplugin": {"width": width, "current_lang": current_lang},
         }
         embedder = Embedder(plugins=plugins, plugin_config=plugin_config)
         url = self.context.external_content_url
         return embedder(url, config={"width": width})
+
+    def has_leadimage(self):
+        if ILeadImage.providedBy(self.context) and getattr(
+            self.context, "image", False
+        ):
+            return True
+        return False
+
+    def image(self):
+        return f"{self.context.absolute_url()}/@@download/image"
 
 
 class EaglebePlugin(Plugin):
@@ -45,8 +77,12 @@ class EllohaPlugin(Plugin):
             #     "ConstellationWidgetContainer": "917c4c52-c997-4077-9135-873a0b2e0c85",
             #     "Idoi": "195ea849-1732-4a69-a051-e7911001cd09"
             # }
+            current_lang = config["current_lang"]
             extra_params = config["extra_params"]
-            error_message = '<div class="elloha">With an elloha plugin, extra params must contain a dictionary with two keys : ConstellationWidgetContainer, Idoi</div>'
+
+            error_message = "With an elloha plugin, extra params must contain a dictionary with two keys : ConstellationWidgetContainer, Idoi"
+            error_message = translate(_(error_message), target_language=current_lang)
+            error_message = f'<div class="elloha elloha_error">{error_message}</div>'
 
             if (
                 extra_params is None
@@ -85,9 +121,16 @@ class EllohaPlugin(Plugin):
 
 class CognitoformPlugin(Plugin):
     def __call__(self, parts, config={}):
-        extra_params = config["extra_params"]
-        error_message = '<div class="cognitoform">With a cognitoform plugin, extra params can be void but if you complete it you must specify : scrolling:(yes/no) and overflow:(hidden/scroll/auto)</div>'
         if "cognitoforms" in parts.netloc:
+            current_lang = config["current_lang"]
+            extra_params = config["extra_params"]
+
+            error_message = "With a cognitoform plugin, extra params can be void but if you complete it you must specify : scrolling:(yes/no) and overflow:(hidden/scroll/auto)"
+            error_message = translate(_(error_message), target_language=current_lang)
+            error_message = (
+                f'<div class="cognitoform cognitoform_error">{error_message}</div>'
+            )
+
             if extra_params is None:
                 return f'<iframe src="{parts.geturl()}" style="border: 0px none; width: {config["width"]}%; overflow: auto;" scrolling="yes"></iframe>'
             else:
@@ -108,6 +151,54 @@ class CognitoformPlugin(Plugin):
         return None
 
 
+class ArcgisPlugin(Plugin):
+    def __call__(self, parts, config={}):
+        if "arcgis" in parts.netloc:
+            # url : https://developers.arcgis.com/
+            # extra_params : {"portal_item_id":"27a432b0835149e6acd3ac39d0e4349c"}
+            current_lang = config["current_lang"]
+            extra_params = config["extra_params"]
+            url = config["url"]
+
+            error_message = "With arcgis plugin, extra params must contain a dictionary with one key : portal_item_id"
+            error_message = translate(_(error_message), target_language=current_lang)
+            error_message = f'<div class="arcgis arcgis_error">{error_message}</div>'
+
+            if (
+                extra_params is None
+                or extra_params[0] != "{"
+                or extra_params[-1] != "}"
+            ):
+                return error_message
+            try:
+                res = json.loads(extra_params.lower())
+            except JSONDecodeError:
+                return error_message
+            if res.get("portal_item_id") is None:
+                return error_message
+
+            portal_item_id = res.get("portal_item_id")
+            msg = "Consult the map"
+            msg = translate(_(msg), target_language=current_lang)
+            return (
+                f'<a href="{url}/view_arcgis?portal_item_id={portal_item_id}">{msg}</a>'
+            )
+        #
+        return None
+
+
 class UnknowServicePlugin(Plugin):
     def __call__(self, parts, config={}):
-        return _("<p class='unknow_service'>Unknow service</p>")
+        current_lang = config["current_lang"]
+        msg = "Unknow service"
+        msg = translate(_(msg), target_language=current_lang)
+        return f'<p class="unknow_service">{msg}</p>'
+
+
+@implementer(IArcgisViewUtils)
+class ArcgisView(BrowserView):
+    _footer = None
+
+    @property
+    def get_portal_item_id(self):
+        return self.request.portal_item_id